scripts/requirements.txt
python-dotenv
boto3
opensearch-py